pyproject.toml
[project]
name = "agent_smith"
version = "0.1.0"
description = "An agent that performs automated tasks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.3",
    "pytest-playwright>=0.7.0",
    "python-dotenv>=1.0.1",
]
